Exercício 30 do curso em video!

+++ ex030.py
@@ -1,4 +1,4 @@
-numero_usuario = float((input("Digite um numero: ")))
+numero_usuario = int((input("Digite um numero: ")))
 num = numero_usuario % 2
 if num == 0:
     print("Esse numero é par!")
